# api/openapi/laserstream-agent-api.yaml
openapi: 3.1.0
info:
  title: LaserStream Agent API
  version: 1.0.0
  description: |
    API surface for AI agents to query verified on-chain analytics, backtest launches,
    configure automations, and exchange signed webhook payloads.
servers:
  - url: http://127.0.0.1:3000
security:
  - ApiKeyAuth: []
paths:
  /api/agent/query:
    post:
      summary: Run natural-language query with guarded SQL generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
      responses:
        "200":
          description: Query completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentQueryResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
  /api/agent/query/report:
    post:
      summary: Strict verified analyst report alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
      responses:
        "200":
          description: Strict report completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentQueryResponse'
  /api/agent/backtest/launches:
    post:
      summary: Backtest historical launches and score predictive signals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchBacktestRequest'
      responses:
        "200":
          description: Backtest output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchBacktestResponse'
  /api/agent/backtest/launches/score:
    post:
      summary: Score a candidate launch from historical signal performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchScoreRequest'
      responses:
        "200":
          description: Candidate score output
          content:
            application/json:
              schema:
                type: object
  /api/agent/webhooks:
    post:
      summary: Register an outbound event webhook destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentWebhookCreateRequest'
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentWebhook'
    get:
      summary: List registered outbound event webhooks
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentWebhook'
  /api/agent/webhooks/inbound:
    post:
      summary: Receive signed inbound webhook payload (replay-protected)
      security: []
      parameters:
        - $ref: '#/components/parameters/WebhookTimestampHeader'
        - $ref: '#/components/parameters/WebhookNonceHeader'
        - $ref: '#/components/parameters/WebhookSignatureHeader'
        - $ref: '#/components/parameters/WebhookKeyIDHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Accepted
        "401":
          description: Signature validation failed
        "409":
          description: Replay nonce detected
    get:
      summary: List inbound webhook records
      responses:
        "200":
          description: Inbound webhook list
          content:
            application/json:
              schema:
                type: object
  /api/agent/automations:
    post:
      summary: Create automation with safety rails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationCreateRequest'
      responses:
        "201":
          description: Automation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'
    get:
      summary: List automations
      responses:
        "200":
          description: Automation list
          content:
            application/json:
              schema:
                type: object
  /api/agent/automations/{id}:
    patch:
      summary: Update automation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationUpdateRequest'
      responses:
        "200":
          description: Automation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'
  /api/agent/automations/{id}/evaluate:
    post:
      summary: Evaluate automation now
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evaluation output
          content:
            application/json:
              schema:
                type: object
  /api/agent/automations/{id}/audit:
    get:
      summary: List automation audit entries
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit records
          content:
            application/json:
              schema:
                type: object
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WebhookTimestampHeader:
      in: header
      name: X-Laser-Timestamp
      required: true
      schema:
        type: string
      description: Unix seconds or RFC3339 timestamp.
    WebhookNonceHeader:
      in: header
      name: X-Laser-Nonce
      required: true
      schema:
        type: string
    WebhookSignatureHeader:
      in: header
      name: X-Laser-Signature
      required: true
      schema:
        type: string
      description: HMAC signature header in format `v1=<hex>`.
    WebhookKeyIDHeader:
      in: header
      name: X-Laser-Key-Id
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AgentQueryRequest:
      type: object
      properties:
        prompt:
          type: string
        execute:
          type: boolean
        mode:
          type: string
          enum: [sql, verified, analyst_report]
        strict_verified:
          type: boolean
        limit:
          type: integer
      required:
        - prompt
    AgentQueryResponse:
      type: object
      properties:
        prompt:
          type: string
        sql:
          type: string
        sql_hash_sha256:
          type: string
        status:
          type: string
        mode:
          type: string
        verified_answer:
          type: string
        evidence_bundle:
          type: object
          additionalProperties: true
    AgentWebhookCreateRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        program_id:
          type: string
        event_type:
          type: string
      required: [url]
    AgentWebhook:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        program_id:
          type: string
        event_type:
          type: string
        is_active:
          type: boolean
    LaunchInput:
      type: object
      properties:
        name:
          type: string
        program_id:
          type: string
        launch_at:
          type: string
          format: date-time
      required: [program_id, launch_at]
    LaunchBacktestThresholds:
      type: object
      properties:
        event_growth_ratio:
          type: number
        unique_signer_growth_ratio:
          type: number
        transaction_growth_ratio:
          type: number
    LaunchBacktestRequest:
      type: object
      properties:
        launches:
          type: array
          items:
            $ref: '#/components/schemas/LaunchInput'
        baseline_hours:
          type: integer
        evaluation_hours:
          type: integer
        min_evaluation_events:
          type: integer
        thresholds:
          $ref: '#/components/schemas/LaunchBacktestThresholds'
      required: [launches]
    LaunchScoreRequest:
      type: object
      properties:
        history_launches:
          type: array
          items:
            $ref: '#/components/schemas/LaunchInput'
        candidate:
          $ref: '#/components/schemas/LaunchInput'
        baseline_hours:
          type: integer
        evaluation_hours:
          type: integer
        min_evaluation_events:
          type: integer
        thresholds:
          $ref: '#/components/schemas/LaunchBacktestThresholds'
      required: [history_launches, candidate]
    LaunchBacktestResponse:
      type: object
      properties:
        status:
          type: string
        config:
          type: object
          additionalProperties: true
        summary:
          type: object
          additionalProperties: true
        launches:
          type: array
          items:
            type: object
            additionalProperties: true
    AutomationCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        program_id:
          type: string
        query_template:
          type: string
        comparator:
          type: string
          enum: [gt, gte, lt, lte]
        threshold:
          type: number
        schedule_minutes:
          type: integer
        window_minutes:
          type: integer
        cooldown_minutes:
          type: integer
        dedupe_minutes:
          type: integer
        retry_max_attempts:
          type: integer
        retry_initial_backoff_ms:
          type: integer
        dry_run:
          type: boolean
        webhook_url:
          type: string
          format: uri
        is_active:
          type: boolean
      required: [name, query_template, comparator, threshold, webhook_url]
    AutomationUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        query_template:
          type: string
        comparator:
          type: string
          enum: [gt, gte, lt, lte]
        threshold:
          type: number
        schedule_minutes:
          type: integer
        window_minutes:
          type: integer
        cooldown_minutes:
          type: integer
        dedupe_minutes:
          type: integer
        retry_max_attempts:
          type: integer
        retry_initial_backoff_ms:
          type: integer
        dry_run:
          type: boolean
        webhook_url:
          type: string
          format: uri
        is_active:
          type: boolean
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        program_id:
          type: string
        query_template:
          type: string
        comparator:
          type: string
        threshold:
          type: number
        schedule_minutes:
          type: integer
        window_minutes:
          type: integer
        cooldown_minutes:
          type: integer
        dedupe_minutes:
          type: integer
        retry_max_attempts:
          type: integer
        retry_initial_backoff_ms:
          type: integer
        dry_run:
          type: boolean
        webhook_url:
          type: string
          format: uri
        is_active:
          type: boolean